requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic<3
SQLAlchemy>=2.0
alembic
faiss-cpu
sentence-transformers
numpy
pymupdf
